Default params friendly fail on DatabaseUpgrader
DatabaseUpgrader will now return an exit code of 0 if you pass it junk
default parameters, in lieu of wrapping the commands.config with Role
checks.
Fortunately, this will be safe in both positive and negative cases,
because if you really need the DbUpgrader to have been passed good
parameters, this will not export the dbUp.bat, in which case the runner
will fail later on.