Fixes a bug when IIS isn't installed yet.
When IIS isn't installed, the IIS version detection code will blow up
because a necessary DLL won't be on the machine yet.  Fortunately, we
only use this check to tell if we need to add a newAppPool or if the app
pool name already exists, and if IIS isn't installed, it definitely
won't exist, so this just wraps that in a try/catch and if it blows up
assumes we need to add the app pool.