Fixes a fetcher problem with odd root folders
If a root folder had any file, even if that file was not an installer,
the service counted that as a good enough folder to stop looking for
folders containing installers.

This only applies to the (currently) deprecated version of the fetcher,
which will need to be reinstated soon to support dev machines.