Fixes an issue with naming collisions.
Ringtail is one of the installers, and all the other installers also
start with Ringtail.
We use 'starts-with' logic because installers have an unpredictable
suffix on them to make them unique.  So this appends the "_" character
to the app name to ensure we only get the ones we really want.