Adds task-level timeouts.
Task level timeouts - default task timeout of 10 minutes.  Can pass in a
different default timeout as a parameter.   Can also have a
"timeout.config" file that allows for per-task whitelisting of timeouts,
with a magic number of '0' for infinite.

Will soon want to add on a 'coure-correct' section in here, so if the
timeout is tripped, we can have a whitelist of a 'cleanup and retry
task'.