Fixes an issue with ServiceInstaller
ServiceInstaller was expecting keys to be named differently than how
Ringtail-Deploy-Web was sending them.

Also - when in this situation, the error handling was super-useless and
required the author to debug the code to figure out what was going
wrong.

I've updated the error handling in this case so that the user has a hope
of understanding what to do.