GenericInstaller auto quote wrapping.
InstallShield requires that all parameters passed in with spaces in them
have to be wrapped in quotes.   Requiring consumers of the service to
know exactly how to wrap their install shield params in quotes causes a
lot of failed builds.

....this adds support for people only using the 2.6.4 API which always
auto-wrapped quotes for them, and people using Ringtail-Deploy-Web which
now helps you with the quotes on the data-entry side.